Interim MAimmunizations improvement. (#50)
* Test for the Mass. error page

As of 11am, it returns a Heroku error page about half the time.
Catch that and fail fast.

Also, check for other CSS selector serach failures, although it's not
clear under what circumstances that might happen.

Previously we got:

  TypeError: Cannot read property 'evaluate' of undefined
      at ScrapeWebsiteData (/Users/jhawk/src/covid-vaccine-scrapers/site-scrapers/MAImmunizations.js:20:48)
      at processTicksAndRejections (internal/process/task_queues.js:93:5)
      at async GetAvailableAppointments (/Users/jhawk/src/covid-vaccine-scrapers/site-scrapers/MAImmunizations.js:5:18)
      at async Promise.all (index 0)
      at async gatherData (/Users/jhawk/src/covid-vaccine-scrapers/main.js:41:19)
      at async execute (/Users/jhawk/src/covid-vaccine-scrapers/main.js:116:2)
      at async /Users/jhawk/src/covid-vaccine-scrapers/main.js:124:3
  The following data would be published:

attempting to iterate over the page navigation.

THe Heroku error page looks like this:
---snip
<!DOCTYPE html><html><head>
		<meta name="viewport" content="width=device-width, initial-scale=1">
		<meta charset="utf-8">
		<title>Application Error</title>
		<style media="screen">
		  html,body,iframe {
			margin: 0;
			padding: 0;
		  }
		  html,body {
			height: 100%;
			overflow: hidden;
		  }
		  iframe {
			width: 100%;
			height: 100%;
			border: 0;
		  }
		</style>
	  </head>
	  <body>
		<iframe src="//www.herokucdn.com/error-pages/application-error.html"></iframe>

	</body></html>
---snip

so we check the page title to find if we're there.

* MAImmunizations.js: Also return false for other failure

Avoid the traceback for every instance of missing page navigation, not
just the Heroku error page.

* Return static copy of MA scrape if we fail

* Fetch the current data.json, use it if neccssary

Note we have a bootstrapping problem, which is why the baseline is
here. Otherwise we can never update ourselves because the retreived
data never has the data to update.

Adds a node-fetch dependancy.