Correctly parse CVS's Mountain timestamp
Unfortunately timezone offsets are not fixed and Mountain time will
change by 1 hour on March 14, which is soon enough that we do care.

Javascript's timezone support is extremely spotty. Without bringing in
a library like moment.js and all the hair that implies, this is about
the best we can do.

To calculate the timezone offset, we render the current datetime as
local time in the two timezones, and then subtract them and convert
from milliseconds to hours, and append the numeric timezone offset
to the returned timestring that is fed to the Date() constructor.

I am so, so, sorry.