CVS.js: Handle reduced JSON
In the past few days CVS's JSON stopped returning numeric availability information
("totalAvailable" and "pctAvailable" keys). This didn't exactly "break" the scraper, but
it causedit to do goofy things like return NaN for availability (but there was none).

Revise the scraper to return a claim of availability as long as CVS
does not say "Fully Booked." Note that in the past, we've seen claims
of boolean availability with claims of "0" numeric availability, and
it's not apparent what that meant. But now we don't really have a choice.

Don't try to parse totalAvailable if it doesn't exist (leads to NaN).

Document the JSON form a bit better in a comment.