CVS.js: Handle reduced JSON (#89)
* CVS.js: Handle reduced JSON

In the past few days CVS's JSON stopped returning numeric availability information
("totalAvailable" and "pctAvailable" keys). This didn't exactly "break" the scraper, but
it causedit to do goofy things like return NaN for availability (but there was none).

Revise the scraper to return a claim of availability as long as CVS
does not say "Fully Booked." Note that in the past, we've seen claims
of boolean availability with claims of "0" numeric availability, and
it's not apparent what that meant. But now we don't really have a choice.

Don't try to parse totalAvailable if it doesn't exist (leads to NaN).

Document the JSON form a bit better in a comment.

* CVS.js: Avoid undefineds: don't use spread (...) syntax

Previously, we took CVS's returned JSON object and added properties to it with
JS's spread (...) syntax in a filter, as:
```
  return { newprop1: newvalue1, newprop2, ... originalObject };
```

Unfortunately, that makes it impossible to suppress `newprop2` if,
e.g., it is `undefined`, as with `totalAvailability` here.

So, we revert to inelegance and instead

```
  r.newprop1 = newvalue1;
  if (newprop2) { r.newprop2 = newprop2 }
  return r;
```

* CVS.js: Cons up a new {}, don't modify CVS's

That is, don't modify the JS object returned by CVS, which includes
fields we don't necessarily use, like `PctAvailable` and `state` (of
the Union) that we don't use.

Instead, return a JS object with only the fields we truly want.

* CVS.js: style: const sted let

* CVS.js: Don't lose TitleCase in site name

Refactoring in 9d81ac inadverently lost TitleCase in the city name as
it appeared in the location -- oops. Cache it in `const city`.