Implement CI workflow using GitHub Actions (#103)
* Start building CI workflow using Github Actions

In an effort to improve code quality of the scrapers, this is a start
of performing continuous integration at development time. Once the
workflow is defined and configured properly, the test suite can be
extended.

* Fix mispelling in workflow name

* Cut unnecessary whitespace

* Make change to os version used to run workflow

GitHub is moving to use ubuntu 20.04 soon. To ensure compatibility
with the AWS Lambda runtime, I set the workflow to remain on 18.04.

* Make workflow trigger on PR against master