Add totalAvailability count to returned object
In this commit, I also refactored GetAvailableAppointments to enable
building a testsuite using Mock and Puppeteer without requiring changes
to the caller. I used the incognitoContext to identify when the test
suite is calling the exported function so that I could intercept
the target url with captured test data in HTML form.