Add Heywood Healthcare (#94)
* Add Heywood Healthcare to sites list

* Add scraper for Heywood Healthcare

This first commit of the scraper includes an initial test indicating
availability based on an alert in the sign-up page.

* Start building CI workflow using Github Actions

In an effort to improve code quality of the scrapers, this is a start
of performing continuous integration at development time. Once the
workflow is defined and configured properly, the test suite can be
extended.

* Make change to os version used to run workflow

GitHub is moving to use ubuntu 20.04 soon. To ensure compatibility
with the AWS Lambda runtime, I set the workflow to remain on 18.04.

* Make workflow trigger on PR against master

* Make directory for Heywood site-scraper

* Cut website field from Heywood JSON

* Refactor site config location

* Refactor GetAvailableAppointments to new standard

* Refactor config location of signUpLink field

* Add logic to handle page without alert div

* Fix case of street address

* Reformat by adding newlines to EOF

* Reformat based on prettier check

* Fix case of street address

* Add totalAvailability count to returned object

In this commit, I also refactored GetAvailableAppointments to enable
building a testsuite using Mock and Puppeteer without requiring changes
to the caller. I used the incognitoContext to identify when the test
suite is calling the exported function so that I could intercept
the target url with captured test data in HTML form.

* Add test suite for Heywood Healthcare

The test suite uses mock, chai, and Puppeteer against the live site,
as well as static HTML captured during development. The test suite
ensures that the scraper handles the documented use cases properly.

The current use cases are the following:
1. Test against the live site
2. Test against static HTML without alert and no availability
3. Test against static HTML with availability

* Add install of chromium-browser

* Add environment variables to GitHub workflow

* Add .prettierignore to exclude static HTML

* Fix live test to support various states

* Add conditional to dateElement