CVSPharmacy: Return time of scrape, not site's time
Alas, for elegance! After discussion at this morning's standup,
conclude that with the current UI, we should return the time we
scrape, rather than the time CVS reports, because it makes our data
looks older than it really is.

Return CVS's timestamp as `siteTimestamp`. Another choice is to return
it in `extraData`, but we're not confident we want to show it now.