Add Baystate Health (#102)
* Add Baystate Health to sites list

* Add scraper for Baystate Health

This first commit of the scraper includes an initial test indicating
availability based on the Registration banner in the sign-up page.

* Make change to directory structure for scraper

The scraper directory structure was improved and this commit is to
bring this scraper inline with the new standard.

* Add site config payload

* Add robustness to scraper

Based on code written for another scraper and feedback from the team,
this commit adds the following:
* Required keys to returned payload
* Ability to be tested using Puppeteer

* Fix signUpLink URL

* Cut unnecessary code

* Add test suite for scraper

* Reformat based on prettier

* Add class names to selector

* Optimize code based on feedback in PR

Made recommended changes to ensure that the scraper was testing against
the intended page. Also incorporated saving the page content to S3, and
sending a message to Slack based on @zpeyton's Walgreens example.

* Optimize test against live site

Based on the changes made in the corresponding scraper, consolidated
the code in the test suite to ensure successful completion.