Removed before() and after() and related requires
Apparently, "test/bootstrap.js" supplies these in the automated test suite. But
note that without require("./bootstrap.js") in this file, it's not possible for the developer to work with this code in VS Code: both "browser" and "expect" will be marked as undefined. I have found no way to insert bootstrap.js into the launch definition in VS Code.